requirements.txt
discord.py>=2.3.0
pandas>=2.0.0
google-cloud-bigquery>=3.10.0
python-dotenv>=1.0.0